+++ Engine/Foundation/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_library(Foundation
         ZeloStringUtil.cpp
+        ZeloIOStream.cpp
         ZeloProfiler.cpp
         )
